Add fabazon, a library for interfacing with AWS through Fabric.
fabazon is designed to make it a bit easier to perform common operations
on S3, ELB, EC2, and other services through a Fabric script.

It makes use of boto, and requires a valid ~/.boto config file, or the
appropriate environment variables.

Currently, fabazon can do the following:

* Upload files to S3, setting their mimetype and public flags.
* Download public or private files from S3 to the remote host, using
  a signed URL.
* Generate and upload directory indexes for directories or whole trees
  on S3.
* Get the EC2 instance ID for the current remote host.
* Attach or detach instances to/from ELB.

This is geared toward the things we need to do for Review Board,
RBCommons, and other related projects, but is intended to be reusable.