Add `#md5=` to file links in directory indexes.
When building directory indexes, we now add a `#md5=` to the end of the
filename in the link, indicating the MD5 checksum of the file, as
provided by S3.

This is used by pip (as part of the package-related PEPs) to verify the
authenticity of the files we provide. This, in theory, allows users to
install without using the --allow-unverified flag (though they still
need --allow-external, for now).

In practice, current versions of pip are still very hostile against
externally-hosted index.html files, which we use. This may be fixed in
pip 1.6, from the sounds of it. If so, we're now better set up for that.

Reviewed at https://reviews.reviewboard.org/r/6394/