Added support for Application Load Balancer management.
This introduces a new class for managing ALBs. The API is different from
that of the old Elastic Load Balancer, and requires a different class.
Consumers can choose which class suits their needs.

This comes with some new utility functions for both types of load
balancers for checking registration and health state and waiting until
instances are marked as healthy or unregistered.

Testing Done:
Made use of this in the RBCommons fabfile to handle deployments.

Reviewed at https://reviews.reviewboard.org/r/9894/