Fixed checking whether an instance is registered on the load balancer.
The API no longer returns a `Reason` field for the health information
for a load balancer target unless there's a value for that field. This
caused a `KeyError` for the check if the target was still registered.

We now check for this in a way that returns the correct value without
failure if the key is not present.